Resolve 9-file conflict set so the Windows build fix (PR #1592) can merge:
- CMakeLists_windows.txt: PR's superset source list (incl. main's new files)
- zinc_cpp CMake: keep main's if(MSVC) block (equivalent to PR's genex)
- zinc main.cpp: PR's manual argv scan (main's getopt_long still breaks MSVC)
- build_windows.cmd: main's vswhere -requires VCTools + PR's -DREPO_ROOT=%REPO_DIR%
  (main's %REPO_ROOT% is undefined in the script -> empty define)
- backend_onnx.cpp: PR's io.h/_access shim below covers both platforms
- q4nx_reader.cpp: PR's <cinttypes>
- gguf_to_onebp.cpp: PR's byte-order guard (drops main's broken static_assert
  referencing possibly-undefined __BYTE_ORDER__), main's PRIu64 formats,
  merged: PRIu64 + PR's if(inf) null guard
- unified_server.cpp: PR's std::filesystem scan + setvbuf _WIN32 guard
  (kept ONE_BIN_DISPATCH comment)

Syntax-checked with g++ -fsyntax-only and mingw cross on all touched files.

+++ engine/fusion/gpu_verify_opt.cpp
@@ -27,10 +27,10 @@ int main() {
     float *d_pk,*d_sc,*d_inorm,*d_pan,*d_qn,*d_kn,*d_fn,*d_hf,*d_xs;
     _Float16 *d_h,*d_qkv,*d_at,*d_ff,*d_ac,*d_res,*d_kc,*d_vc; int8_t *d_i8;
     auto ml=[&](auto&p_,size_t b){hipMalloc((void**)&p_,b);};
-    ml(d_pk,L*per_layer*4);ml(d_sc,L*per_sc*4);
-    ml(d_inorm,L*H*4);ml(d_pan,L*H*4);ml(d_qn,L*HD*4);ml(d_kn,L*HD*4);
+    ml(d_pk,(size_t)L*per_layer*4);ml(d_sc,(size_t)L*per_sc*4);
+    ml(d_inorm,(size_t)L*H*4);ml(d_pan,(size_t)L*H*4);ml(d_qn,(size_t)L*HD*4);ml(d_kn,(size_t)L*HD*4);
     ml(d_fn,H*4);ml(d_hf,8192*4);ml(d_h,H*2);
-    ml(d_qkv,(NH*HD+2*NKV*HD)*2);ml(d_at,NH*HD*2);
+    ml(d_qkv,(NH*HD+2*NKV*HD)*2);ml(d_at,(size_t)NH*HD*2);
     ml(d_ff,2*IM*2);ml(d_ac,IM*2);ml(d_res,H*2);
     ml(d_i8,H);ml(d_xs,4);
     int MP=4096;
@@ -96,7 +96,7 @@ int main() {
             for(int h=0;h<NKV;h++)rcpp_rmsnorm_fp16(d_qkv+NH*HD+h*HD,d_kn+l*HD,d_qkv+NH*HD+h*HD,1e-6f,HD,s);
             rcpp_rope_fp16(d_qkv+NH*HD,l,1000000.0f,NKV,HD,s);
             
-            size_t kvo=(size_t)l*MP*NKV*HD+l*NKV*HD;
+            size_t kvo=(size_t)l*MP*NKV*HD+(size_t)l*NKV*HD;
             hipMemcpy(d_kc+kvo,d_qkv+NH*HD,NKV*HD*2,hipMemcpyDeviceToDevice);
             hipMemcpy(d_vc+kvo,d_qkv+NH*HD+NKV*HD,NKV*HD*2,hipMemcpyDeviceToDevice);
